+++ nextstep-di/src/main/java/nextstep/di/factory/domain/factory/SingleInstanceRegistry.java
